refactor(012): split analysis-table.ts into focused units (US3)
Decompose the 637-line analysis-table.ts (now 301) into:
- services/analysis-display.ts: pure groupEntriesByCell/sortByTimestamp + CellEntry
- enhancers/analysis-persistence.ts: handleCellEdit/saveCellData
- enhancers/analysis-instructor-overlay.ts: createStudentEntriesDisplay, show/hide entries, getCurrentPageId

Add StorageService.updateRecordWithAnalysis (mirrors updateRecordWithAnswer)
so the persistence path no longer hand-builds analysis records. analysis-table.ts
retains only lifecycle/orchestration. Re-point test imports; add a focused
analysis-display unit test.
